+++ CupheadOnline/Plugin.cs
@@ -186,6 +186,6 @@ namespace CupheadOnline
     {
         public const string GUID    = "com.cupheadonline.mod";
         public const string NAME    = "CupheadOnline";
-        public const string VERSION = "1.1.0";
+        public const string VERSION = "1.1.1";
     }
 }